Rental Controller Refactor. Use Services for better code readability.
