+++ Boardium/Boardium/Program.cs
@@ -25,6 +25,7 @@ builder.Services.AddTransient<QrCodeService>();
 builder.Services.AddTransient<PickupCodeGenerator>();
 builder.Services.AddTransient<DateBetweenChecker>();
 builder.Services.AddTransient<GameService>();
+builder.Services.AddTransient<RentalService>();
 builder.Services.AddSingleton(typeof(IConverter), new SynchronizedConverter(new PdfTools()));
 builder.Services.Configure<EmailSettings>(builder.Configuration.GetSection("SmtpSettings"));
 builder.Services.AddTransient<EmailService>();
